refactor(*): remove `@Valid` from response DTOs
