--- src/main/java/sleppynavigators/studyupbackend/presentation/authentication/dto/TokenResponse.java
@@ -1,7 +1,0 @@
-package sleppynavigators.studyupbackend.presentation.authentication.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record TokenResponse(@NotBlank String accessToken,
-                            @NotBlank String refreshToken) {
-}
