+++ src/main/java/sleppynavigators/studyupbackend/presentation/authentication/dto/response/TokenResponse.java
@@ -1,0 +1,4 @@
+package sleppynavigators.studyupbackend.presentation.authentication.dto.response;
+
+public record TokenResponse(String accessToken, String refreshToken) {
+}
