--- src/main/java/sleppynavigators/studyupbackend/presentation/group/dto/GroupInvitationResponse.java
@@ -1,6 +1,0 @@
-package sleppynavigators.studyupbackend.presentation.group.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record GroupInvitationResponse(@NotBlank String invitationId) {
-}
