+++ src/main/java/sleppynavigators/studyupbackend/presentation/group/dto/response/GroupInvitationResponse.java
@@ -1,0 +1,4 @@
+package sleppynavigators.studyupbackend.presentation.group.dto.response;
+
+public record GroupInvitationResponse(String invitationId) {
+}
